move mock_data.py out of cached data dir
